Proj8_Arxiv_Pipeline/extraction/requirements.txt
requests
pandas
pyarrow
azure-storage-blob